ml/training/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml_pipeline"
version = "0.1.0"
description = "Biotrace training pipeline and inference utilities"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.18",
  "opencv-python>=4.8.0",
  "xgboost>=2.1.0",
  "joblib>=1.3.0",
  "mlflow-skinny>=2.14.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0"
]

[project.scripts]
ml-export-dataset = "ml_pipeline.export_dataset:main"
ml-train = "ml_pipeline.train:main"
ml-run = "ml_pipeline.run_pipeline:main"

[tool.setuptools]
packages = ["ml_pipeline"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
